Syntax Commit 5
Main function added. Main opens the file and if the file is empty, an error message is displayed. Otherwise, it calls the getChar function to access the first lexeme on the line. It then calls the lex function which processes the lexeme which will later determine what to do based on the character type. Then, the expr function is called to determine the parsing sequence.